Adding support for 12bit MCP320x Three new sub-classes, read method now appears in all sub-classes, constructor only in parent. AnalogIn.value uses calculated shift values based on new class variable BITS. #41
